cleanup

+++ src/cli.ts
@@ -16,9 +16,12 @@ const args = process.argv.slice(2);
 
 let serviceName = "";
 let command = "";
+let verbose = false;
 for (let i = 0; i < args.length; i++) {
   const arg = args[i];
-  if (arg === 'install' || arg === '/install' || arg === '-install' || arg === 'i' || arg === '/i' || arg === '-i') {
+  if (arg === '-v' || arg === '-verbose' || arg === '/v' || arg === '/verbose') {
+    verbose = true;
+  } else if (arg === 'install' || arg === '/install' || arg === '-install' || arg === 'i' || arg === '/i' || arg === '-i') {
     command = 'install';
   } else if (arg === 'uninstall' || arg === '/uninstall' || arg === '-uninstall' || arg === 'u' || arg === '/u' || arg === '-u') {
     command = 'uninstall';
@@ -33,12 +36,21 @@ for (let i = 0; i < args.length; i++) {
   }
 }
 
+function Run(cmd: string) {
+  try {
+    const output = childProcess.execSync(cmd).toString();
+    if(verbose) console.log(output);
+    return output
+  } catch (error) {
+    return error.toString();
+  }
+}
 function installService(svcName: string, serviceName: string, script: string): void {
   let scriptPath = path.join(__dirname, script);
-  if(!fs.existsSync(scriptPath)) {
+  if (!fs.existsSync(scriptPath)) {
     scriptPath = path.join(__dirname, "dist", script);
   }
-  if(!fs.existsSync(scriptPath)) {
+  if (!fs.existsSync(scriptPath)) {
     console.log("Failed locating " + script)
     return;
   }
@@ -48,17 +60,15 @@ function installService(svcName: string, serviceName: string, script: string): v
     const svcPath = `node ${scriptPath}`;
     const svcRegPath = `HKLM\\SYSTEM\\CurrentControlSet\\Services\\${svcName}`;
 
-    const cmd = `sc create "${svcName}" binPath="${svcPath}" DisplayName="${serviceName}" Description="${svcDescription}" start=auto`;
-    const output = childProcess.execSync(cmd).toString();
-    console.log(output);
-
-    const regCmd = `reg add ${svcRegPath} /v DependOnService /t REG_MULTI_SZ /d EventLog /f`;
-    const regOutput = childProcess.execSync(regCmd).toString();
-    console.log(regOutput);
-
+    Run(`sc create "${svcName}" binPath="${svcPath}" DisplayName="${serviceName}" Description="${svcDescription}" start=auto`)
+    Run(`reg add ${svcRegPath} /v DependOnService /t REG_MULTI_SZ /d EventLog /f`)
+    Run(`net start ${svcName}`)
     console.log(`Service "${serviceName}" installed successfully.`);
   } else {
     const svcPath = `/etc/systemd/system/${svcName}.service`;
+    if(fs.existsSync(svcPath)) {
+      UninstallService(svcName, serviceName);
+    }
     const svcContent = `
       [Unit]
       Description=${serviceName}
@@ -73,63 +83,38 @@ function installService(svcName: string, serviceName: string, script: string): v
       [Install]
       WantedBy=multi-user.target
    `;
-
     fs.writeFileSync(svcPath, svcContent);
-    console.log(`Service file created at "${svcPath}".`);
-
-    const cmd = `systemctl enable ${serviceName}.service`;
-    const output = childProcess.execSync(cmd).toString();
-    console.log(output);
-
-    try {
-      const cmd2 = `systemctl start ${serviceName}.service`;
-      const output2 = childProcess.execSync(cmd2).toString();
-      console.log(output2);
-    } catch (error) {
-    }
-
+    if(verbose) console.log(`Service file created at "${svcPath}".`);
+    Run(`systemctl enable ${serviceName}.service`)
+    Run(`systemctl start ${serviceName}.service`)
 
     console.log(`Service "${serviceName}" installed successfully.`);
-    console.log(`to validate use\nsudo systemctl status ${svcName}.service\nsudo journalctl -efu ${svcName}`)
+    console.log(`sudo systemctl status ${svcName}.service\nsudo journalctl -efu ${svcName}`)
   }
 }
 
-function UninstallService(svcName:string, serviceName:string):void {
+function UninstallService(svcName: string, serviceName: string): void {
   if (os.platform() === 'win32') {
     const svcRegPath = `HKLM\\SYSTEM\\CurrentControlSet\\Services\\${svcName}`;
 
-    const cmd = `sc delete "${svcName}"`;
-    const output = childProcess.execSync(cmd).toString();
-    console.log(output);
-
-    const regCmd = `reg delete ${svcRegPath} /f`;
-    const regOutput = childProcess.execSync(regCmd).toString();
-    console.log(regOutput);
-
+    Run(`net stop ${svcName}`)
+    Run(`sc delete "${svcName}"`)
+    Run(`reg delete ${svcRegPath} /f`)
     console.log(`Service "${serviceName}" uninstalled successfully.`);
   } else {
     const svcPath = `/etc/systemd/system/${svcName}.service`;
-
-    try {
-      const cmd2 = `systemctl stop ${serviceName}.service`;
-      const output2 = childProcess.execSync(cmd2).toString();
-      console.log(output2);
-    } catch (error) {
+    if(fs.existsSync(svcPath)) {
+      Run(`systemctl stop ${serviceName}.service`)
+      Run(`systemctl disable ${serviceName}.service`)
+      fs.unlinkSync(svcPath);
+      console.log(`Service file removed at "${svcPath}".`);
     }
-
-    const cmd = `systemctl disable ${serviceName}.service`;
-    const output = childProcess.execSync(cmd).toString();
-    console.log(output);
-
-    fs.unlinkSync(svcPath);
-    console.log(`Service file removed at "${svcPath}".`);
-
     console.log(`Service "${serviceName}" uninstalled successfully.`);
   }
 }
 
 async function main() {
-  if(serviceName == "" || serviceName == null) {
+  if (serviceName == "" || serviceName == null) {
     serviceName = "nodeagent"
   }
   if (command === 'install' || command == "") {
@@ -140,18 +125,18 @@ async function main() {
       assistentConfig = require(path.join(os.homedir(), ".openiap", "config.json"));
     }
     assistentConfig.apiurl = prompts(`apiurl (Enter for ${assistentConfig.apiurl})? `, assistentConfig.apiurl);
-    if(assistentConfig.apiurl == null) process.exit(0);
+    if (assistentConfig.apiurl == null) process.exit(0);
 
-    if(assistentConfig.jwt != null && assistentConfig.jwt != "") {
-      var reuse = prompts(`Reuse existing token (Enter for yes)? (yes/no) `, {value: "yes", autocomplete: ()=> ["yes", "no"]});
-      if(reuse == null) process.exit(0);
-      if(reuse == "no") {
+    if (assistentConfig.jwt != null && assistentConfig.jwt != "") {
+      var reuse = prompts(`Reuse existing token (Enter for yes)? (yes/no) `, { value: "yes", autocomplete: () => ["yes", "no"] });
+      if (reuse == null) process.exit(0);
+      if (reuse == "no") {
         assistentConfig.jwt = "";
       }
-  
+
     }
 
-    if(assistentConfig.jwt == null || assistentConfig.jwt == "") {
+    if (assistentConfig.jwt == null || assistentConfig.jwt == "") {
       const [tokenkey, signinurl] = await agenttools.AddRequestToken(assistentConfig.apiurl)
       console.log(`Please open ${signinurl} in your browser and login with your OpenIAP account`)
       const jwt = await agenttools.WaitForToken(assistentConfig.apiurl, tokenkey);
@@ -159,7 +144,7 @@ async function main() {
     }
     if (!fs.existsSync(path.join(os.homedir(), ".openiap"))) fs.mkdirSync(path.join(os.homedir(), ".openiap"));
     fs.writeFileSync(path.join(os.homedir(), ".openiap", "config.json"), JSON.stringify(assistentConfig));
-  
+
     console.log(`Installing service "${serviceName}"...`);
     installService(serviceName, serviceName, 'agent.js');
   } else if (command === 'uninstall') {
